Populate files with content from get_file function calls
This is the client-side implementation of the get_file HOT
intrinsic function. The data for each resource is recursively
crawled to look for {get_file: <path>} calls. The content for
the URL specified by get_file is fetched and the content stored
in the files field in the API request.

If a relative path is specified, that path is converted to an
absolute URL, and that URL is used as the key for the files
content entry. The template is also modified to replace the
relative path with the absolute URL.

This implements the client-side portion of
blueprint get-file

Change-Id: If1da98578bb2919e3c8325442224376d3fadb0bc