Make POST/PUT redirects use original methods
Commands like 'stack-update' and 'action-suspend' use PUT or
POST methods when interacting with the Heat API. They also
require the use of redirects. Currently redirects for these use
GET instead of the original method which leads to undesired
outcomes. This commit ensures that the redirect always uses the
original method.

Change-Id: Iee578ea23c842e48f2d73992804f9c2efb9c7f43
Closes-bug: #1280158